syntax: interpret 'text' blocks as Rust code
Whenever I use `text` code fence blocks in a doc comment, the syntax
highlighter appears to get deeply confused. It results in all of
syntax highlighting below the `text` block being off. Interestingly,
unadorned or `ignore` code fence blocks are highlighted correctly.
Presumably because there is something different in the syntax definition
for code fence blocks that are meant to be highlighted as Rust.

In any case, I couldn't figure out how to just fix non-Rust code fence
blocks because this definition is absolute insanity. So I instead just
made `text` be highlighted as Rust code. Not ideal, but much better than
what I was dealing with. This doesn't fix other types of fence blocks,
but in practice, `text` is the only non-Rust block I use.